load books, people and rentals data preservation
